+++ Cards/pptoll_cfg.py
@@ -6,7 +6,7 @@ from Config.Integration.vegas_cff import integrator
 from Config.PDG_cfi import PDG, registerParticle
 
 #--- example of an auxiliary particles definition
-#registerParticle(1000001, 'sd_l', mass=100., charge=1., fermion=True)
+#registerParticle(1000001, 'sd_l', mass=100., charge=1., fermion=True) # right now, only fermionic coupling handled
 
 process = kt.process.clone('pptoll',
     processParameters = cepgen.Parameters(