+++ CepGen/IO/GenericExportHandler.cpp
@@ -13,58 +13,61 @@
 
 namespace cepgen
 {
-  namespace output
+  namespace io
   {
-    GenericExportHandler::GenericExportHandler() :
-      event_num_( 0. )
+    GenericExportHandler::GenericExportHandler( const std::string& name ) :
+      name_( name ), event_num_( 0. )
+    {}
+
+    GenericExportHandler::~GenericExportHandler()
     {}
 
     std::string
-    GenericExportHandler::banner( const Parameters& params )
+    GenericExportHandler::banner( const Parameters& params, const std::string& prep )
     {
       std::ostringstream os;
       os
-        << "  ***** Sample generated with CepGen v" << version() << " *****\n"
-        << "  * process: " << params.processName() << " (" << params.process()->mode() << ")\n"
-        << "  * beams:\n"
-        << "  *   - " << params.kinematics.incoming_beams.first << "\n"
-        << "  *   - " << params.kinematics.incoming_beams.second << "\n";
+        << prep << "  ***** Sample generated with CepGen v" << version() << " *****\n"
+        << prep << "  * process: " << params.processName() << " (" << params.process()->mode() << ")\n"
+        << prep << "  * beams:\n"
+        << prep << "  *   - " << params.kinematics.incoming_beams.first << "\n"
+        << prep << "  *   - " << params.kinematics.incoming_beams.second << "\n";
       if ( params.process()->mode() != KinematicsMode::ElasticElastic && !params.hadroniserName().empty() )
-        os << "  * hadroniser: " << params.hadroniserName() << "\n";
+        os << prep << "  * hadroniser: " << params.hadroniserName() << "\n";
       os
-        << "  *--- incoming state\n";
+        << prep << "  *--- incoming state\n";
       if ( params.kinematics.cuts.initial.q2.valid() )
         os
-          << "  * Q2 range (GeV2): "
+          << prep << "  * Q2 range (GeV2): "
           << params.kinematics.cuts.initial.q2 << "\n";
       if ( params.process()->mode() != KinematicsMode::ElasticElastic
         && params.kinematics.cuts.remnants.mass_single.valid() )
         os
-          << "  * remnants mass range (GeV/c2): "
+          << prep << "  * remnants mass range (GeV/c2): "
           << params.kinematics.cuts.remnants.mass_single << "\n";
-      os << "  *--- central system\n";
+      os << prep << "  *--- central system\n";
       if ( params.kinematics.cuts.central.pt_single.valid() )
         os
-          << "  * single particle pt (GeV/c): "
+          << prep << "  * single particle pt (GeV/c): "
           << params.kinematics.cuts.central.pt_single << "\n";
       if ( params.kinematics.cuts.central.energy_single.valid() )
         os
-          << "  * single particle energy (GeV): "
+          << prep << "  * single particle energy (GeV): "
           << params.kinematics.cuts.central.energy_single << "\n";
       if ( params.kinematics.cuts.central.eta_single.valid() )
         os
-          << "  * single particle eta: "
+          << prep << "  * single particle eta: "
           << params.kinematics.cuts.central.eta_single << "\n";
       if ( params.kinematics.cuts.central.pt_sum.valid() )
         os
-          << "  * total pt (GeV/c): "
+          << prep << "  * total pt (GeV/c): "
           << params.kinematics.cuts.central.mass_sum << "\n";
       if ( params.kinematics.cuts.central.mass_sum.valid() )
         os
-          << "  * total invariant mass (GeV/c2): "
+          << prep << "  * total invariant mass (GeV/c2): "
           << params.kinematics.cuts.central.mass_sum << "\n";
       os
-        << "  **************************************************";
+        << prep << "  **************************************************";
       return os.str();
     }
   }